Rework the file structure (#107)
* Rework the file structure

* Bump version
