+++ package.json
@@ -1,6 +1,6 @@
 {
     "name": "fitracker",
-    "version": "1.1.0",
+    "version": "1.1.1",
     "main": "expo-router/entry",
     "scripts": {
         "start": "cross-env NODE_ENV=development expo start --go",
