--- src/app/(app)/(profile)/_layout.tsx
@@ -1,6 +1,0 @@
-import { Slot } from 'expo-router';
-import React from 'react';
-
-export default function ProfileLayout() {
-    return <Slot />;
-}
