+++ src/app/(app)/_layout.tsx
@@ -1,17 +1,12 @@
 import IonIcons from '@expo/vector-icons/Ionicons';
 import { Tabs } from 'expo-router';
 import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import PopoverMenu from 'src/components/common/PopoverMenu';
-import { useAuth } from 'src/context/auth-context/AuthContext';
+import { useDispatch } from 'react-redux';
 import { useLocalStorage } from 'src/hooks/common/useLocalStorage';
-import { RootState } from 'src/redux/Store';
 import { WORKOUT_FORM_STORAGE_KEY } from 'src/redux/workout-form/WorkoutFormMiddleware';
 import { loadWorkoutOnRender } from 'src/redux/workout-form/WorkoutFormSlice';
-import { useTheme, View } from 'tamagui';
 
 export default function AppLayout() {
-    const username = useSelector((state: RootState) => state.user.username);
     const dispatch = useDispatch();
     const { getFromStorage } = useLocalStorage();
 
@@ -26,11 +21,12 @@ export default function AppLayout() {
                 console.error(e);
             });
     }, []);
+
     return (
         <>
             <Tabs>
                 <Tabs.Screen
-                    name='(workout-tracker)'
+                    name='workout-tracker'
                     options={{
                         headerShown: false,
                         tabBarLabel: 'Start Workout',
@@ -40,10 +36,9 @@ export default function AppLayout() {
                     }}
                 />
                 <Tabs.Screen
-                    name='(profile)'
+                    name='profile'
                     options={{
-                        title: username,
-                        headerRight: ProfileScreenHeaderRightComponent,
+                        headerShown: false,
                         tabBarLabel: 'Profile',
                         tabBarIcon: ({ color, size }) => (
                             <IonIcons name='person-outline' color={color} size={size} />
@@ -52,25 +47,5 @@ export default function AppLayout() {
                 />
             </Tabs>
         </>
-    );
-}
-
-function ProfileScreenHeaderRightComponent() {
-    const theme = useTheme();
-    const { logout } = useAuth();
-
-    return (
-        <View marginRight='$space.4'>
-            <PopoverMenu
-                menuOptions={[
-                    {
-                        text: 'Logout',
-                        icon: 'log-out',
-                        iconColor: theme.background.val,
-                        action: logout,
-                    },
-                ]}
-            />
-        </View>
     );
 }
